+++ package.json
@@ -7,7 +7,8 @@
     "build": "vue-cli-service build",
     "build-staging": "vue-cli-service build --mode staging",
     "test:unit": "vue-cli-service test:unit --coverageProvider=v8",
-    "lint": "vue-cli-service lint"
+    "lint": "vue-cli-service lint",
+    "lint:fix": "vue-cli-service lint --fix"
   },
   "dependencies": {
     "@types/mathjax": "^0.0.37",
